Route tests changed to single file with an array
