+++ server/routes/type.js
@@ -1,0 +1,9 @@
+module.exports = {
+  method: 'GET',
+  path: '/type',
+  options: {
+    handler: {
+      view: 'type'
+    }
+  }
+}
