+++ server/views/type.njk
@@ -1,0 +1,7 @@
+{% extends 'layout.njk' %}
+
+{% block content %}
+  <div class="govuk-body">
+    <h1>test</h1>
+  </div>
+{% endblock %}
