+++ test/happy-test.js
@@ -3,7 +3,7 @@ const Code = require('code')
 const lab = exports.lab = Lab.script()
 const createServer = require('../server')
 
-lab.experiment('Route tests', () => {
+lab.experiment('Happy path Test', () => {
   let server
 
   // Create server before the tests
@@ -11,38 +11,6 @@ lab.experiment('Route tests', () => {
     server = await createServer()
   })
   // Happy tests
-  lab.test('GET / route works', async () => {
-    const options = {
-      method: 'GET',
-      url: '/'
-    }
-
-    const response = await server.inject(options)
-    Code.expect(response.statusCode).to.equal(200)
-    Code.expect(response.headers['content-type']).to.include('text/html')
-  })
-
-  lab.test('GET /about route works', async () => {
-    const options = {
-      method: 'GET',
-      url: '/about'
-    }
-
-    const response = await server.inject(options)
-    Code.expect(response.statusCode).to.equal(200)
-    Code.expect(response.headers['content-type']).to.include('text/html')
-  })
-
-  lab.test('GET /authority route works', async () => {
-    const options = {
-      method: 'GET',
-      url: '/authority'
-    }
-
-    const response = await server.inject(options)
-    Code.expect(response.statusCode).to.equal(200)
-    Code.expect(response.headers['content-type']).to.include('text/html')
-  })
 
   lab.test('POST /authority route works', async () => {
     const options = {
@@ -57,26 +25,16 @@ lab.experiment('Route tests', () => {
     Code.expect(response.headers['content-type']).to.include('text/html')
   })
 
-  lab.test('GET /type route works', async () => {
-    const options = {
-      method: 'GET',
-      url: '/type'
-    }
-    const response = await server.inject(options)
-    Code.expect(response.statusCode).to.equal(200)
-    Code.expect(response.headers['content-type']).to.include('text/html')
-  })
+  // lab.test('POST /type route works', async () => {
+  //   const options = {
+  //     method: 'POST',
+  //     url: '/type',
+  //     payload: { type: 'recovery' }
+  //   }
 
-  lab.test('POST /type route works', async () => {
-    const options = {
-      method: 'POST',
-      url: '/type',
-      payload: { type: 'recovery' }
-    }
-
-    const response = await server.inject(options)
-    Code.expect(response.statusCode).to.equal(302)
-    Code.expect(response.headers['location']).to.equal('./')
-    Code.expect(response.headers['content-type']).to.include('text/html')
-  })
+  //   const response = await server.inject(options)
+  //   Code.expect(response.statusCode).to.equal(302)
+  //   Code.expect(response.headers['location']).to.equal('./type')
+  //   Code.expect(response.headers['content-type']).to.include('text/html')
+  // })
 })
